fix(iris): cast MTLRenderStages long to int for STAGE_ALL
MTLRenderStages.VertexAndFragment.value is `long` (the enum field is
`public final long value`), but STAGE_ALL was declared `int`. Add an
explicit (int) cast to fix the "incompatible types: possible lossy
conversion from long to int" compile error that broke CI run 29984193914.

This was the only compile error blocking the M4b+M4c Metal rendering
implementation (fullscreen triangle + final pass).

+++ src/main/java/com/metallum/client/metal/render/MetalIrisRenderer.java
@@ -46,7 +46,7 @@ public final class MetalIrisRenderer {
     private static final Logger LOGGER = LoggerFactory.getLogger("MetalUniversal");
 
     /** Stage mask: vertex + fragment (used for buffer/texture binding). */
-    private static final int STAGE_ALL = MTLRenderStages.VertexAndFragment.value;
+    private static final int STAGE_ALL = (int) MTLRenderStages.VertexAndFragment.value;
 
     /** Maximum number of texture slots to bind dummy textures to. */
     private static final int MAX_TEXTURE_SLOTS = 8;
